// src/app/models/personalData.ts
export interface PersonalData {
    firstName: string;
    lastName: string;
    email: string;
    telephone: number | null;
    address: string;
    zipCode: number | null;
    city: string;
    birthDay: string;
    birthMonth: string;
    birthYear: string;
    citizenship: string;
    gender: string;
    civilState: string;
    linkedIn: string;
    website: string;
}